lexer patched back together
