+++ src/toolchain/mod.rs
@@ -1,2 +1,3 @@
-pub mod cursor;
+//pub mod diagnostics;
 pub mod lexer;
+//pub mod parser;
